+++ climate_categories/__init__.py
@@ -5,7 +5,7 @@ names of categorizations. To access the example categorization `Excat`, simply u
 
 __author__ = """Mika Pflüger"""
 __email__ = "mika.pflueger@climate-resource.com"
-__version__ = "0.10.1"
+__version__ = "0.10.6"
 
 import importlib
 import importlib.resources
@@ -54,6 +54,7 @@ RCMIP = _read_py_hier("RCMIP")
 gas = _read_py_hier("gas")
 ISO3 = _read_py_hier("ISO3")
 ISO3_GCAM = _read_py_hier("ISO3_GCAM")
+FAO = _read_py_hier("FAO")
 
 
 def find_code(code: str) -> set[Category]: